backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Werkzeug==3.1.3
APScheduler==3.10.4
